feat(e2e): add EncryptionKeyBootstrap component for key initialization and public key management
